First commit of Teller Class on new branch.
Teller Class will replace Bank class as the driver class which is user-facing. Adding a simple login GUI to make it neater and give some structure other than running in the terminal. User will now have to login with appropriate userID and userPassword field in the Account Class (to be implemented). Bank Class will now hold an ArrayList of Accounts to allow for multiple accounts.