Added Welcome and CreateAccount classes.
More organizational changes as I learn more. I decided to have the Teller class just be a driver class for the program and so Welcome holds all of the welcome page code and CreateAccount holds that page. Began redoing the Bank class code as well to update it with the ArrayList functionality. It's all kind of a mess right now, but still opens the initial window. Functionality is coming soon.